Various improvements on the ConflictReadGraph.

+++ src/ConflictReadGraph.cpp
@@ -138,7 +138,7 @@ void ConflictReadGraph::colorConnectedComponent(const vector<VertexId>& componen
         //     " color " << color0 << endl;
         ConflictReadGraphVertex& vertex0 = getVertex(u0);
         SHASTA_ASSERT(vertex0.clusterId = ConflictReadGraphVertex::invalid);
-        vertex0.clusterId = color0;
+        vertex0.clusterId = uint32_t(color0);
 
         // Increment adjacentColoredCount for adjacent uncolored vertices.
         for(EdgeId edgeId: incidentEdges(u0)) {
@@ -190,7 +190,7 @@ void ConflictReadGraph::colorConnectedComponent(const vector<VertexId>& componen
 
 
     // Find the number of colors used.
-    Int maxColor = 0;
+    uint32_t maxColor = 0;
     for(Int v=0; v<n; v++) {
         maxColor = max(maxColor, getVertex(component[v]).clusterId);
     }
